application/requirements.txt
Flask
python-dotenv
awscli
boto3
jira
markdown
requests